+++ Compose-Flight-Search-App/core/data/src/main/kotlin/com/example/flightsearchapp/core/data/model/FavoriteWithAirportsConverter.kt
@@ -1,6 +1,7 @@
 package com.example.flightsearchapp.core.data.model
 
 import com.example.flightsearchapp.core.database.model.FavoriteWithAirports
+import com.example.flightsearchapp.core.model.Flight
 
 fun FavoriteWithAirports.asFlight(): Flight =
     Flight(
