+++ Compose-Flight-Search-App/core/data/src/main/kotlin/com/example/flightsearchapp/core/data/repository/AllFlightsRepository.kt
@@ -1,7 +1,7 @@
 package com.example.flightsearchapp.core.data.repository
 
 import com.example.flightsearchapp.core.data.di.DispatcherDefault
-import com.example.flightsearchapp.core.data.model.Flight
+import com.example.flightsearchapp.core.model.Flight
 import com.example.flightsearchapp.core.data.model.asFlight
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -21,7 +21,7 @@ class AllFlightsRepository @Inject constructor(
     @DispatcherDefault private val defaultDispatcher: CoroutineDispatcher,
 ) {
     @OptIn(ExperimentalCoroutinesApi::class)
-    fun getAllFlightsStream(departureId: Long): Flow<List<Flight>> {
+    fun getAllFlightsStream(departureId: Long): Flow<List<com.example.flightsearchapp.core.model.Flight>> {
         return airportsRepository.getAirportNullByIdStream(airportId = departureId)
             .flatMapLatest { departureAirport ->
                 if (departureAirport == null) {
@@ -32,7 +32,7 @@ class AllFlightsRepository @Inject constructor(
                         favoritesRepository.getFavoritesStream(),
                     ) { arriveAirports, favorites ->
                         withContext(defaultDispatcher) {
-                            val flightsModel = mutableListOf<Flight>()
+                            val flightsModel = mutableListOf<com.example.flightsearchapp.core.model.Flight>()
 
                             val favoriteHashMap: HashMap<String, MutableSet<String>> = hashMapOf()
                             favorites.forEach { favorite ->
@@ -47,7 +47,7 @@ class AllFlightsRepository @Inject constructor(
 
                             arriveAirports.forEach { arriveAirport ->
                                 flightsModel.add(
-                                    Flight(
+                                    com.example.flightsearchapp.core.model.Flight(
                                         id = "${departureAirport.id}_${arriveAirport.id}",
                                         departureIataCode = departureAirport.iataCode,
                                         departureName = departureAirport.name,
@@ -69,7 +69,7 @@ class AllFlightsRepository @Inject constructor(
     }
 
     @OptIn(ExperimentalCoroutinesApi::class)
-    fun getAllFavoriteFlightsStream(): Flow<List<Flight>> =
+    fun getAllFavoriteFlightsStream(): Flow<List<com.example.flightsearchapp.core.model.Flight>> =
         favoritesRepository.getFavoriteWithAirports().mapLatest { favoriteWithAirports ->
             favoriteWithAirports.map { favoriteWithAirport ->
                 favoriteWithAirport.asFlight()
